Rules to live by.
Add your friends for more fun!
Create game invites and don't leave people out.
Play for fun or get competitive!
Find a friend to play with,  or make a new one it's up to you just don't have a bad time. We are all friends here girls and guys alike, or what ever sex you classify as. If you want to classify as a giant bear wearing a ballerina costume go for it! If you want to be a cat woman, or man we can dig it, or if you want to be a space alien go for it! (I kind of like that last one myself). Just be yourself and we will accept you. Now get out there and have some fun!!!!!!